Extract export file deletion to separate utility function
Extract export file deletion to separate utility function and refactor its usages
