+++ cypress/e2e/timetablePassingTimes.cy.ts
@@ -36,6 +36,7 @@ import {
   removeFromDbHelper,
   SupportedResources,
 } from '../utils';
+import { deleteExportFile } from './utils';
 
 // These infralink IDs exist in the 'infraLinks.sql' test data file.
 // These form a straight line on Eerikinkatu in Helsinki.
@@ -283,11 +284,7 @@ describe('Timetable import and export', () => {
   });
 
   afterEach(() => {
-    const exportDate = DateTime.now().toISODate();
-    const exportFilePath = `${Cypress.config(
-      'downloadsFolder',
-    )}/jore4-export-${exportDate}.csv`;
-    cy.task('deleteFile', exportFilePath);
+    deleteExportFile();
 
     removeFromDbHelper(dbResources);
     cy.task('truncateTimetablesDatabase');
