+++ cypress/e2e/timetableValidityPeriod.cy.ts
@@ -36,6 +36,7 @@ import {
   insertToDbHelper,
   removeFromDbHelper,
 } from '../utils';
+import { deleteExportFile } from './utils';
 
 // These infralink IDs exist in the 'infraLinks.sql' test data file.
 // These form a straight line on Eerikinkatu in Helsinki.
@@ -286,12 +287,7 @@ describe('Timetable validity period', () => {
   });
 
   afterEach(() => {
-    const exportDate = DateTime.now().toISODate();
-    const exportFilePath = `${Cypress.config(
-      'downloadsFolder',
-    )}/jore4-export-${exportDate}.csv`;
-    cy.task('deleteFile', exportFilePath);
-
+    deleteExportFile();
     removeFromDbHelper(dbResources);
     cy.task('truncateTimetablesDatabase');
   });
