+++ cypress/e2e/utils/file.ts
@@ -1,0 +1,9 @@
+import { DateTime } from 'luxon';
+
+export const deleteExportFile = () => {
+  const exportDate = DateTime.now().toISODate();
+  const exportFilePath = `${Cypress.config(
+    'downloadsFolder',
+  )}/jore4-export-${exportDate}.csv`;
+  cy.task('deleteFile', exportFilePath);
+};
