+++ cypress/e2e/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './db-utils';
+export * from './file';
 export * from './string';
 export * from './tasks';
